Added a pause button and redid the way the control buttons go.
Now there's three control buttons.  The Go button, obviously, goes.  It
also resumes, in a pinch.  When recording, there's a Pause button and no
Stop button.  When paused, there's a Go button (to resume) and a Stop
button (to stop).  Later, stopping will return to the main menu.

Also, the screen logging is now standardized.

Note to self: I can't serialize Threads, apparently.  Must rethink how the
AssemblyLine works such that I can pass a reference to it back in the
event of onSaveInstanceState.  Maybe it'll become a Service.

git-svn-id: https://drivelapse.googlecode.com/svn/trunk@11 4d9c8ec4-2505-d0a5-56dc-5841f2f03469