Fixed state saving in the main DriveLapse class.
PictureTaker is no longer serializable.  In fact, it really wasn't that way
to begin with, owing to Context not being serializable.  Since it isn't
necessary for the same object to be active between config changes or
whatnot, it can now simply be recreated with the same timestamp, meaning
it will use the same directory as need be.

Also, I don't have any idea what was going on with the state saving in the
main DriveLapse activity.  Now it actually does something.

git-svn-id: https://drivelapse.googlecode.com/svn/trunk@19 4d9c8ec4-2505-d0a5-56dc-5841f2f03469